+++ doc/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, Marco Dal Molin, Dmitri Kavetski, Fabrizio Fenicia'
 author = 'Marco Dal Molin, Dmitri Kavetski, Fabrizio Fenicia'
 
 # The full version, including alpha/beta/rc tags
-release = '1.1.0-rc'
+release = '1.2.0'
 
 
 # -- General configuration ---------------------------------------------------